Ajuste nome da pasta e navegacao sidebar
